pyproject.toml
[tool.poetry]
name = "requests-query"
version = "0.1.0"
description = ""
authors = ["Dmitry Nekrasov <demonekr@yandex.ru>"]
readme = "README.md"
packages = [{include = "requests_query"}]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = {extras = ["all"], version = "^0.99.1"}
fake-headers = "^1.0.2"
ruff = "^0.0.276"
pytest-asyncio = "^0.21.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
